pyproject.toml
[tool.poetry]
name = "afc-dut-test"
version = "0.1.0"
description = "AFC DUT Test"
authors = ["Loki <yongxiang.zhan@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.0"
pytest-html = "^4.1.0"
pytest-xdist = "^3.3.0"
loguru = "^0.7.0"
pyyaml = "^6.0.1"
paramiko = "^3.5.0"
allure-pytest = "^2.13.5"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = "."
addopts = """
    --alluredir=./reports/allure-results
    -v
"""
